+++ src/main/java/edu/harvard/iq/dataverse/provenance/ProvEntityFileData.java
@@ -1,4 +1,4 @@
-package edu.harvard.iq.dataverse;
+package edu.harvard.iq.dataverse.provenance;
 /**
  *
  * @author madunlap
